POLIO-1947: cleanup & organize lqas folder
