+++ plugins/polio/js/src/domains/LQAS-IM/LQAS/CountryOverview/LqasCountryDataView.tsx
@@ -1,18 +1,19 @@
-import { Campaign, MapShapes, Side } from '../../../..//constants/types';
+import { Campaign, MapShapes, Side } from '../../../../constants/types';
 import React, { FunctionComponent, useMemo } from 'react';
 import { ConvertedLqasImData, LqasImMapLayer } from '../../types';
 import { baseUrls } from '../../../../constants/urls';
 import { useLqasImTabState } from '../../shared/Tabs/useLqasImTabState';
 import { useMapShapes } from '../../shared/hooks/api/useMapShapes';
 import { getLqasMapLayer, useLqasMapHeaderData } from './utils';
-import { LqasCountryMapV2 } from './LqasCountryMapV2';
+
 import { Divider, Paper } from '@mui/material';
 import { LIST, MAP, paperElevation } from '../../shared/constants';
 import { LqasImMapHeader } from '../../shared/Map/LqasImMapHeader';
 import { LqasSummary } from './LqasSummary';
 import { LqasImTabs } from '../../shared/Tabs/LqasImTabs';
 import { DropdownOptions } from 'Iaso/types/utils';
-import { LqasCountryListOverview } from './LqasCountryListOverview';
+import { LqasCountryListView } from './LqasCountryListView';
+import { LqasCountryMapView } from './LqasCountryMapView';
 
 type Props = {
     roundNumber?: number;
@@ -24,11 +25,11 @@ type Props = {
     debugData?: Record<string, unknown> | null;
     params: Record<string, string | undefined>;
     isFetching: boolean;
-    roundOptions: DropdownOptions<number>[];
+    roundOptions?: DropdownOptions<string>[];
 };
 
 const baseUrl = baseUrls.lqasCountry;
-export const LqasCountryContainer: FunctionComponent<Props> = ({
+export const LqasCountryDataView: FunctionComponent<Props> = ({
     params,
     side,
     countryId,
@@ -83,7 +84,7 @@ export const LqasCountryContainer: FunctionComponent<Props> = ({
             <Divider />
             <LqasImTabs tab={tab} handleChangeTab={handleChangeTab} />
             {tab === MAP && (
-                <LqasCountryMapV2
+                <LqasCountryMapView
                     side={side}
                     roundNumber={roundNumber}
                     campaign={campaign}
@@ -98,7 +99,8 @@ export const LqasCountryContainer: FunctionComponent<Props> = ({
                 />
             )}
             {tab === LIST && (
-                <LqasCountryListOverview
+                <LqasCountryListView
+                    side={side}
                     shapes={mainLayer}
                     regionShapes={regionShapes}
                     isFetching={
